# Add a performance-optimized version of length for Cofree