--- a/src/Control/Comonad/Cofree.hs
+++ b/src/Control/Comonad/Cofree.hs
@@ -192,6 +192,8 @@ instance Foldable f => Foldable (Cofree f) where
   foldMap f = go where
     go (a :< as) = f a `mappend` foldMap go as
   {-# INLINE foldMap #-}
+  length = go 0 where
+    go s (_ :< as) = foldl' go (s + 1) as
 
 instance Foldable1 f => Foldable1 (Cofree f) where
   foldMap1 f = go where