@glguy 2015-09-26
More CPP needed for old base